+++ src/client/main.tsx
@@ -1,4 +1,5 @@
 import { StrictMode } from 'react';
+import './animations.css';
 import { createRoot } from 'react-dom/client';
 import { App } from './App';